%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{finalize.Job}
\alias{finalize.Job}
\alias{Job.finalize}
\alias{finalize.Job}
\alias{finalize,Job-method}


\title{Finalizes job}

\description{
 Finalizes job by unlocking it, if locked.
}

\usage{\method{finalize}{Job}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\seealso{
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
