%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getOutputPath.Job}
\alias{getOutputPath.Job}
\alias{Job.getOutputPath}
\alias{getOutputPath.Job}
\alias{getOutputPath,Job-method}


 \title{Gets the output path of the job}

 \description{
  Gets the output path of the job.
  It is recommended to write logs, store results etc. to this directory.
 }

 \usage{\method{getOutputPath}{Job}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
  Returns a \code{\link[base]{character}} string.
 }

 \seealso{
   For more information see \code{\link{Job}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
