%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getPath.Job}
\alias{getPath.Job}
\alias{Job.getPath}
\alias{getPath.Job}
\alias{getPath,Job-method}

 \alias{Job.setPath}
\alias{setPath.Job}
\alias{setPath,Job-method}


 \title{Gets the path to the job}

 \description{
  Gets the path to the job.
 }

 \usage{\method{getPath}{Job}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
  Returns a \code{\link[base]{character}} string.
 }

 \seealso{
   For more information see \code{\link{Job}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
