%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  JobBatch.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{getRoot.JobBatch}
\alias{getRoot.JobBatch}
\alias{JobBatch.getRoot}
\alias{getRoot.JobBatch}
\alias{getRoot,JobBatch-method}

 \alias{JobBatch.setRoot}
\alias{setRoot.JobBatch}
\alias{setRoot,JobBatch-method}


 \title{Gets the root path of the job batch}

 \description{
  Gets the root path of the job batch.
 }

 \usage{\method{getRoot}{JobBatch}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
  Returns a \code{\link[base]{character}} string.
 }

 \seealso{
   For more information see \code{\link{JobBatch}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
