%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  JobBatch.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{getSummary.JobBatch}
\alias{getSummary.JobBatch}
\alias{JobBatch.getSummary}
\alias{getSummary.JobBatch}
\alias{getSummary,JobBatch-method}


\title{Gets a summary of the jobs directory}

\description{
 Gets a summary of the jobs directory.
}

\usage{\method{getSummary}{JobBatch}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{list}}.
}

\seealso{
  For more information see \code{\link{JobBatch}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
