%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hasFailed.Job}
\alias{hasFailed.Job}
\alias{Job.hasFailed}
\alias{hasFailed.Job}
\alias{hasFailed,Job-method}


\title{Checks if the job failed}

\description{
 Checks if the job failed either during its main loop or duing it finalization.
}

\usage{\method{hasFailed}{Job}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the job failed, otherwise \code{\link[base:logical]{FALSE}}.
}

\seealso{
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
