%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{hasStoredImage.Job}
\alias{hasStoredImage.Job}
\alias{Job.hasStoredImage}
\alias{hasStoredImage.Job}
\alias{hasStoredImage,Job-method}


\title{Checks if a stored job image exists}

\description{
 Checks if a stored job image exists.
}

\usage{\method{hasStoredImage}{Job}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if a stored image exists, otherwise \code{\link[base:logical]{FALSE}}.
}

\seealso{
  \code{\link[R.batch:saveImage.Job]{*saveImage}()}.
  \code{\link[R.batch:loadStoredImage.Job]{*loadStoredImage}()}.
  \code{\link[R.batch:removeStoredImage.Job]{*removeStoredImage}()}.
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
