%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{isStarted.Job}
\alias{isStarted.Job}
\alias{Job.isStarted}
\alias{isStarted.Job}
\alias{isStarted,Job-method}


\title{Checks if the job is started or not}

\description{
 Checks if the job is started or not.
}

\usage{\method{isStarted}{Job}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the job is started, otherwise \code{\link[base:logical]{FALSE}}.
}

\seealso{
  \code{\link[R.batch:run.Job]{*run}()}.
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
