%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{loadStoredImage.Job}
\alias{loadStoredImage.Job}
\alias{Job.loadStoredImage}
\alias{loadStoredImage.Job}
\alias{loadStoredImage,Job-method}


\title{Reload a stored job image}

\description{
 Reload a stored job image, if exists.
}

\usage{\method{loadStoredImage}{Job}(this, cleanup=TRUE, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns a \code{\link[base]{character}} \code{\link[base]{vector}} with the names of the reloaded fields.
}

\seealso{
  \code{\link[R.batch:hasStoredImage.Job]{*hasStoredImage}()}.
  \code{\link[R.batch:removeStoredImage.Job]{*removeStoredImage}()}.
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
