%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


 % The Job class
 \alias{addPersistentField}
 \alias{addPersistentField.Job}
 \alias{backupLog}
 \alias{backupLog.Job}
 \alias{getDependencies}
 \alias{getDependencies.Job}
 \alias{getDependenciesRaw}
 \alias{getDependenciesRaw.Job}
 \alias{getField}
 \alias{getField.Job}
 \alias{getFigurePath.Job}
 \alias{getLog}
 \alias{getLog.Job}
 \alias{getLogPath.Job}
 \alias{getPersistentPath}
 \alias{getPersistentPath.Job}
 \alias{getResultPath.Job}
 \alias{getRoot.Job}
 \alias{isErroneous}
 \alias{isExisting}
 \alias{isSinked}
 \alias{isFinished}
 \alias{isFinished.Job}
 \alias{isNewer}
 \alias{isNewer.Job}
 \alias{isPersistentField}
 \alias{isPersistentField.Job}
 \alias{lastModified}
 \alias{lastModified.Job}
 \alias{listDir}
 \alias{listDir.Job}
 \alias{listFiles}
 \alias{listFiles.Job}
 \alias{moveOutputFilesTo}
 \alias{moveOutputFilesTo.Job}
 \alias{resetToRun}
 \alias{resetToRun.Job}
 \alias{setField}
 \alias{setField.Job}
 \alias{setStatus}
 \alias{touch}
 \alias{touch.Job}
 \alias{untouch}
 \alias{untouch.Job}
 \alias{hasStoredImage}
 \alias{loadStoredImage}
 \alias{removeStoredImage}
 \alias{resetLog}
 \alias{setLabel}
 \alias{setFigurePath}
 \alias{setFigurePath.Job}
 \alias{setOutputPath}
 \alias{setOutputPath.Job}
 \alias{setPath}
 \alias{hasFailed}
 \alias{isDone}
 \alias{isLocked}
 \alias{isStarted}
 \alias{lock}
 \alias{run}
 \alias{saveImage}
 \alias{setup}
 \alias{showWarnings}
 \alias{sink}
 \alias{sink.default}
 \alias{sourceDirectoryWithPreprocessor}
 \alias{sourceDirectoryWithPreprocessor.Job}
 \alias{sourceHotCode}
 \alias{unlock}
 \alias{unsink}
 \alias{wasInterrupted}
 \alias{wasSuccessful}
 \alias{writeToLog}

 % The JobBatch class
 \alias{checkRequirements}
 \alias{clean}
 \alias{copyFrom}
 \alias{createStub}
 \alias{findJobs}
 \alias{resetJobs}
 \alias{getDirectory}
 \alias{getErroneousPath}
 \alias{getFailedPath}
 \alias{getFigurePath}
 \alias{getFinishedPath}
 \alias{getInputPath}
 \alias{getInterruptedPath}
 \alias{getLogPath}
 \alias{getNextJob}
 \alias{getOutputPath}
 \alias{getResultPath}
 \alias{getRoot}
 \alias{getRunningPath}
 \alias{getSrcPath}
 \alias{getSummary}
 \alias{getTodoPath}
 \alias{main}
 \alias{moveJobTo}
 \alias{setRoot}
 \alias{validate}
 \alias{getLabel}
 \alias{getRunAndFinishJob}
 \alias{getStatus}
 \alias{getSubdirectory}
 \alias{validate}
 \alias{setupDemo}

 \description{
   This page contains aliases for all "non-documented" objects that
   \code{R CMD check} detects in this package.

   Almost all of them are \emph{generic} functions that have specific
   document for the corresponding method coupled to a specific class.
   Other functions are re-defined by \code{setMethodS3()} to
   \emph{default} methods. Neither of these two classes are non-documented
   in reality.
   Some methods may also be deprecated and are therefore removed from the
   help pages.

   However, because this package is in alpha or beta version, some
   methods are still to be documented. The help for such methods are
   linked to this page.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
