%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  JobBatch.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{print.JobBatch}
\alias{print.JobBatch}
\alias{JobBatch.print}
\alias{print.JobBatch}
\alias{print,JobBatch-method}


\title{Prints a summary of the jobs directory}

\description{
 Prints a summary of the jobs directory.
}

\usage{\method{print}{JobBatch}(x, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\seealso{
  For more information see \code{\link{JobBatch}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
