%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{removeStoredImage.Job}
\alias{removeStoredImage.Job}
\alias{Job.removeStoredImage}
\alias{removeStoredImage.Job}
\alias{removeStoredImage,Job-method}


\title{Removes stored job image}

\description{
 Removes stored job image, if exists.
}

\usage{\method{removeStoredImage}{Job}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\seealso{
  \code{\link[R.batch:saveImage.Job]{*saveImage}()}.
  \code{\link[R.batch:hasStoredImage.Job]{*hasStoredImage}()}.
  \code{\link[R.batch:loadStoredImage.Job]{*loadStoredImage}()}.
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
