%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{saveImage.Job}
\alias{saveImage.Job}
\alias{Job.saveImage}
\alias{saveImage.Job}
\alias{saveImage,Job-method}


\title{Save an image of the job}

\description{
 Save an image of the job.
}

\usage{\method{saveImage}{Job}(this, label=getLabel(this), ...)}

\arguments{
 \item{label}{A \code{\link[base]{character}} string to be used in the filename.}
 \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\seealso{
  \code{\link[R.batch:hasStoredImage.Job]{*hasStoredImage}()}.
  \code{\link[R.batch:loadStoredImage.Job]{*loadStoredImage}()}.
  \code{\link[R.batch:removeStoredImage.Job]{*removeStoredImage}()}.
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
