%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setLabel.Job}
\alias{setLabel.Job}
\alias{Job.setLabel}
\alias{setLabel.Job}
\alias{setLabel,Job-method}


\title{Sets the label of the job}

\description{
 Sets the label of the job.
}

\usage{\method{setLabel}{Job}(this, label=NULL, ...)}

\arguments{
 \item{label}{A \code{\link[base]{character}} string. If \code{\link[base]{NULL}}, label is reset.}
 \item{...}{Not used.}
}

\value{
 Returns nothing.
}

\seealso{
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
