%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{setStatus.Job}
\alias{setStatus.Job}
\alias{Job.setStatus}
\alias{setStatus.Job}
\alias{setStatus,Job-method}


\title{Sets the status of the job}

\description{
 Sets the status of the job.
}

\usage{\method{setStatus}{Job}(this, status, ...)}

\arguments{
 \item{status}{A \code{\link[base]{character}} string.}
 \item{...}{Not used.}
}

\value{
 Returns invisibly the previous status as a \code{\link[base]{character}} string.
}

\seealso{
  \code{\link[R.batch:getStatus.Job]{*getStatus}()}.
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
