%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{showWarnings.Job}
\alias{showWarnings.Job}
\alias{Job.showWarnings}
\alias{showWarnings.Job}
\alias{showWarnings,Job-method}


 \title{Displays warning generated while running this job}

 \description{
   Displays warning generated while running this job.
 }

 \usage{\method{showWarnings}{Job}(this, reset=TRUE, show=TRUE, ...)}

 \arguments{
  \item{reset}{If \code{\link[base:logical]{TRUE}}, warnings are removed after being viewed,
    otherwise not.}
  \item{show}{If \code{\link[base:logical]{TRUE}}, warnings are displayed, otherwise not.}
  \item{...}{Not used.}
 }

 \value{
   Returns nothing.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

 \seealso{
   For more information see \code{\link{Job}}.
 }
\keyword{internal}
\keyword{methods}
