%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  JobBatch.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{validate.JobBatch}
\alias{validate.JobBatch}
\alias{JobBatch.validate}
\alias{validate.JobBatch}
\alias{validate,JobBatch-method}


 \title{Validates the job batch}

 \description{
  Validates the job batch by asserting that all required subdirectories exist.
  An exception is thrown if not valid.
 }

 \usage{\method{validate}{JobBatch}(this, ...)}

 \arguments{
  \item{...}{Not used.}
 }

 \value{
  Returns nothing.
 }

 \seealso{
   For more information see \code{\link{JobBatch}}.
 }

 \author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
