%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{wasInterrupted.Job}
\alias{wasInterrupted.Job}
\alias{Job.wasInterrupted}
\alias{wasInterrupted.Job}
\alias{wasInterrupted,Job-method}


\title{Checks if the job was interrupted}

\description{
 Checks if the job was interrupted either during its main loop or duing it finalization.
}

\usage{\method{wasInterrupted}{Job}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the job was interrupted, otherwise \code{\link[base:logical]{FALSE}}.
}

\seealso{
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
