%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{wasSuccessful.Job}
\alias{wasSuccessful.Job}
\alias{Job.wasSuccessful}
\alias{wasSuccessful.Job}
\alias{wasSuccessful,Job-method}


\title{Checks if the job was completed successfully}

\description{
 Checks if the job was completed successfully.
}

\usage{\method{wasSuccessful}{Job}(this, ...)}

\arguments{
 \item{...}{Not used.}
}

\value{
 Returns \code{\link[base:logical]{TRUE}} if the job was successful, otherwise \code{\link[base:logical]{FALSE}}.
}

\seealso{
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
