%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  Job.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{writeToLog.Job}
\alias{writeToLog.Job}
\alias{Job.writeToLog}
\alias{writeToLog.Job}
\alias{writeToLog,Job-method}


\title{Writes to log file}

\description{
 Writes to log file.
}

\usage{\method{writeToLog}{Job}(this, ..., collapse=" ")}

\arguments{
 \item{...}{Arguments passed to the \code{\link[R.utils]{cat.Verbose}} method.}
 \item{collapse}{A \code{\link[base]{character}} string to concatenate objects in a list.}
}

\value{
 Returns nothing.
}

\seealso{
  \code{\link[R.batch:resetLog.Job]{*resetLog}()}.
  For more information see \code{\link{Job}}.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{internal}
\keyword{methods}
\keyword{programming}
